tests: Split localed wrapper from module
